refactor: Extract to client.rs
lib.rs has become too large, so extract to client.rs.
No breaking changes.
